==> notebooks/02_process_input.ipynb <==
## Imports

Importing the required modules for the project.

Pandas is used for its advanced dataframe functionality.

import pandas as pd

## Processing the Input

We will work with the file called `parameters.xlsx`, which contains the parameters to be used for our study. We will merge the different sheets inside of it in order to produce a combined Excel file, which will be used in `get_responses.ipynb` to run our trials and get responses from the OpenAI API.

Reading in the initial Excel sheet.

excel_sheets = pd.read_excel(
    "../input/parameters/01_parameters.xlsx",
    sheet_name=None
    )

for sheet_name, df in excel_sheets.items():
    print(sheet_name)
    display(df.head(3))

Combining the different tables present in the Excel file using the primary and foreign keys.

merge_map = [
    {
        "left": "tbl_studies",
        "right": "tbl_scenarios",
        "on": "study_id"
    },

    {
        "right": "tbl_stories",
        "on": "story_id"
    },

    {
        "right": "tbl_chats",
        "on": "chat_id"
    },

    {
        "right": "tbl_system_messages",
        "left_on": ["study_id", "chat_language"],
        "right_on": ["study_id", "system_message_language"]
    }
]

def merge_sheets(excel_sheets, merge_map: list[dict[str, any]]):
    previous_df = None

    for map in merge_map:
        left_df = None
        if "left" in map:
            left_df = excel_sheets[map["left"]]
            left_suffix = f"_{map["left"].split("_")[-1]}"
        else:
            left_df = previous_df
            left_suffix = f"{left_suffix}{right_suffix}"

        right_df = excel_sheets[map["right"]]
        right_suffix = f"_{map["right"].split("_")[-1]}"

        if "on" in map:
            combined_df = pd.merge(
                left_df,
                right_df,
                how="left",
                on=map["on"],
                suffixes=(left_suffix, right_suffix)
            )
        elif "left_on" in map and "right_on" in map:
            combined_df = pd.merge(
                left_df,
                right_df,
                how="left",
                left_on=map["left_on"],
                right_on=map["right_on"],
                suffixes=(left_suffix, right_suffix)
            )
        
        previous_df = combined_df
    
    return previous_df

combined_df = merge_sheets(excel_sheets, merge_map)

combined_df = pd.merge(
    excel_sheets["tbl_studies"],
    excel_sheets["tbl_scenarios"],
    how="left",
    on="study_id",
    suffixes=("_studies", "_scenarios")
    )

combined_df = pd.merge(
    combined_df,
    excel_sheets["tbl_stories"],
    how="left",
    on="story_id",
    suffixes=("_studies_scenarios", "_stories")
)

combined_df = pd.merge(
    combined_df,
    excel_sheets["tbl_chats"],
    how="left",
    on="chat_id",
    suffixes=("_studies_scenarios_stories", "_chats")
)

combined_df = pd.merge(
    combined_df,
    excel_sheets["tbl_system_messages"],
    how="left",
    left_on=["study_id", "chat_language"],
    right_on=["study_id", "system_message_language"],
    suffixes=("_studies_scenarios_stories_chats", "_system_messages")
)

Let us have a look at the combined dataframe's properties.

combined_df.info()

Let us preview the first 3 rows of our dataframe.

combined_df.head(3)

Right now, each row in our dataframe represents a scenario. The questions to be asked in each scenario can be found in the `questions` column. This column is currently of the type `string`. More specifically, it's a list of comma separated values. 

What we want to do now is split them into a list, and expand our dataframe so that each row represent a question to be asked in a given scenario.

Splitting the questions.

def split_into_columns(series: pd.Series):
    series.str.split(", ", expand=True)

questions_expanded = combined_df["questions"].str.split(", ", expand=True)

questions_expanded.info()

questions_expanded.head(3)

combined_df = pd.melt(
    pd.concat(
        [combined_df, questions_expanded],
        axis=1
        ),
        id_vars=combined_df.columns,
        value_vars=questions_expanded.columns,
        var_name="question_index",
        value_name="question_id"
        )

combined_df = pd.merge(
    combined_df,
    excel_sheets["tbl_questions"],
    on="question_id").dropna(subset=["question_id", "question_content"])

combined_df = pd.merge(
    combined_df,
    excel_sheets["tbl_answers"],
    how="left",
    on=["story_id", "question_id"],
    suffixes=("_studies_scenarios_stories_chats_messages", "_system_messages")
)

combined_df.info()

combined_df.columns

Now, the amount of rows represents the amount of questions to be asked.

combined_df = combined_df.sort_values(
    by=["study_id", "scenario_id", "chat_id", "question_index"]
    )

combined_df.head(3)

combined_df[combined_df["scenario_id"] == 61]

combined_df[combined_df["scenario_id"] == 31]

We will now save it to a `.csv` file, which will be used to get the desired responses from the OpenAI API.

combined_df.to_csv("../input/parameters_expanded/03_parameters_expanded.csv")
