==> notebooks/06_combine_responses.ipynb <==
import pandas as pd
import glob
from pathlib import Path
import datetime

directory = Path("../output/responses_raw/")

combined_df = pd.concat(
    (pd.read_csv(file, index_col=0) for file in directory.glob("*.csv")),
    ignore_index=True
)

combined_df.head(3)

current_datetime = datetime.datetime.now().strftime("%Y-%m-%dT%H-%M-%S")
current_datetime

combined_df.to_csv(
    f"../output/responses_combined/07_combined_responses.csv"
    )

combined_df.groupby(["scenario_id", "question_id"]).size().mean() == 20
