==> notebooks/08_extract_unique_answers.ipynb <==
import pandas as pd
import glob

file_pattern = "../output/responses_combined/07_combined_responses_*.csv"

matching_files = glob.glob(file_pattern)

timestamps = [datetime.strptime(file.split("_")[-1].split(".")[0], "%Y-%m-%dT%H-%M-%S") for file in matching_files]
timestamps.sort(reverse=True)

#latest_file = next(file for file in matching_files if timestamps[0].strftime("%Y-%m-%dT%H-%M-%S") in file)
latest_file = "../output/responses_combined/07_combined_responses.csv"

df = pd.read_csv(
    latest_file,
    index_col=0
)

df.head(3)

[column for column in df.columns if "logprobs" not in column]

columns = [
    "story_id",
    "story_content",
    "question_id",
    "question_content",
    "answer_correct",
    "choices_0_message_content",
    "question_options"
]

unique_answers = df[columns].drop_duplicates()

unique_answers.to_csv(
    f"../output/responses_unique_answers/09_unique_answers.csv"
    )
