==> notebooks/11_code_responses.ipynb <==
import pandas as pd
import glob
from datetime import datetime

coded_responses_path = "../input/responses_unique_answers_coded/10_unique_answers_coded.csv"

coded_df = pd.read_csv(
    coded_responses_path,
    index_col=0
)

all_responses_path = "../output/responses_combined/07_combined_responses.csv"

responses_df = pd.read_csv(
    all_responses_path,
    index_col=0
)

coded_df.info()

responses_df.info()

combined_df = pd.merge(
    left=responses_df,
    right=coded_df,
    on=["story_id", "story_content", "question_id",
        "question_content", "answer_correct", "choices_0_message_content"
        ]
)

combined_df.info()

combined_df = combined_df.drop(columns=["logprob_sum", "logprob_mean", "logprob_sum_exp", "perplexity_score"])

import numpy as np

combined_df["logprob_sum"] = combined_df["choices_0_logprobs_content_0_logprob"]
combined_df["logprob_mean"] = combined_df["choices_0_logprobs_content_0_logprob"]
combined_df["logprob_sum_exp"] = np.exp(combined_df["logprob_sum"])
combined_df["perplexity_score"] = np.exp(-combined_df["logprob_mean"])

combined_df.info()

combined_df.to_csv(
    "../output/responses_coded/12_coded_responses.csv"
)
