==> notebooks/13_prepare_for_analysis.ipynb <==
import pandas as pd
import glob
from datetime import datetime

If you don't have to manually code responses (i.e. you set `with_options=True` while fetching responses), uncomment the second line below.

If you have to manually code responses (for example, with open-ended questions), then you should run notebook 08 first and extract unique answers, then code said answers, place them in `../input/responses_unique_answers_coded/10_unique_answers_coded.csv`, then run notebook 11 to combine the coded answers with the original set of responses. Then, uncomment the second line below and run this notebook.

#path = "../output/responses_coded/12_coded_responses.csv"
path = "../output/responses_combined/07_combined_responses.csv"

df = pd.read_csv(
    path,
    index_col=0
)

def create_new_column(row):
    language = row['story_language']
    
    if pd.notna(row['question_has_fbv_zan']):
        return f"{language}-zan"
    elif pd.notna(row['question_has_fbv_san']):
        return f"{language}-san"
    else:
        return f"{language}"

df["custom_question_language_category"] = df.apply(create_new_column, axis=1)

df.to_csv(path)

", ".join(list(df.columns))

print(df.iloc[38]["choices_0_message_content"])
